# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PromptHub API Documentation",
        "description": "API completa para gerenciamento de prompts de IA e agentes AI. Sistema que permite autenticacao via email/senha ou Google OAuth, cadastro e gerenciamento de agentes de IA personalizados, envio de prompts para OpenAI e Google Gemini.",
        "contact": {
            "name": "PromptHub Support",
            "email": "suporte@prompthub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://dflix-prompthub-backend.a8zera.easypanel.host/api",
            "description": "Production Server (Easypanel)"
        },
        {
            "url": "https://api-prompthub.marcioleite.cloud/api",
            "description": "Production Server (Alternative)"
        }
    ],
    "paths": {
...
        "/tasks": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Criar nova tarefa de IA",
                "description": "Cria uma tarefa de processamento de IA em background. A tarefa e enviada para uma fila e processada de forma assincrona. Retorna imediatamente com status PENDING.",
                "operationId": "createTask",
                "requestBody": {
                    "description": "Dados da tarefa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prompt",
                                    "ai_api_key"
                                ],
                                "properties": {
                                    "prompt": {
                                        "description": "Prompt para o agente de IA (maximo 2000 caracteres)",
                                        "type": "string",
                                        "maxLength": 2000,
                                        "example": "Analise este texto e extraia os principais topicos..."
                                    },
                                    "ai_api_key": {
                                        "description": "Chave de API do provedor de IA (OpenAI, Gemini, etc)",
                                        "type": "string",
                                        "example": "sk-proj-abc123..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Tarefa aceita e enviada para processamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa enviada para o agente de IA."
                                        },
                                        "task_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "PENDING",
                                                "PROCESSING",
                                                "COMPLETED",
                                                "FAILED"
                                            ],
                                            "example": "PENDING"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validacao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The prompt field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/{task}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obter status de uma tarefa",
                "description": "Retorna os detalhes e o status atual de uma tarefa de processamento de IA. O usuario so pode visualizar suas proprias tarefas.",
                "operationId": "getTask",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "UUID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "prompt": {
                                            "type": "string",
                                            "example": "Analise este texto..."
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "PENDING",
                                                "PROCESSING",
                                                "COMPLETED",
                                                "FAILED"
                                            ],
                                            "example": "COMPLETED"
                                        },
                                        "result": {
                                            "description": "Resultado do processamento quando status=COMPLETED",
                                            "type": "string",
                                            "example": "Analise completa: Os principais topicos sao...",
                                            "nullable": true
                                        },
                                        "error": {
                                            "description": "Mensagem de erro quando status=FAILED",
                                            "type": "string",
                                            "example": "API timeout after 30s",
                                            "nullable": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-25T10:00:00.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-25T10:00:30.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuario nao autorizado a ver esta tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nao autorizado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa nao encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Task]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Nao autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}